requirements.txt
spacy>=3.7.0
PyYAML>=6.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.0
tqdm>=4.65.0
reportlab>=4.0.0
openai>=1.0.0
scipy>=1.10.0
requests>=2.28.0
flask>=3.0.0
flask-cors>=4.0.0
openpyxl>=3.0.0
umap-learn>=0.5.0
sentence-transformers>=2.2.0
pyarrow>=12.0.0
scikit-learn>=1.2.0
# optional
wordcloud>=1.9.0
symspellpy>=6.7.0
hdbscan>=0.8.0
